pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pi_inventory_system"
version = "0.1.0"
description = "Voice-controlled inventory management system for a chest freezer using Raspberry Pi"
authors = [
    {name = "Mike Cialowicz"}
]
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "SpeechRecognition==3.10.0",
    "pyaudio==0.2.14",
    "spacy==3.8.0",
    "Pillow==10.2.0",
    "pyttsx3>=2.90",
    "playsound==1.2.2",
    "word2number>=1.1",
    "en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz",
]

[project.optional-dependencies]
test = [
    "pytest==8.0.0",
]
raspberry-pi = [
    "RPi.GPIO>=0.7.1",
    "inky>=1.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"] 